+++ crates/bevy_math/src/primitives/mod.rs
@@ -6,6 +6,14 @@ mod dim2;
 pub use dim2::*;
 mod dim3;
 pub use dim3::*;
+mod frustum;
+pub use frustum::Frustum;
+mod half_space;
+pub use half_space::HalfSpace;
+/// Clean me up!
+/// Code adopted from bevy_render.
+/// Needs to be public to allow bevy_render registery types from here.
+pub mod legacy_bevy_render;
 #[cfg(feature = "serialize")]
 mod serde;
